+++ hutool-core/src/test/java/cn/hutool/core/collection/CollUtilTest.java
@@ -533,17 +533,34 @@ public class CollUtilTest {
 		// Assert result
 		Assert.assertNull(retval);
 	}
-	
+
 	@Test
 	public void sortPageAllTest() {
-		ArrayList<Integer> list = CollUtil.newArrayList(1,2,3,4,5,6,7,8,9);
+		ArrayList<Integer> list = CollUtil.newArrayList(1, 2, 3, 4, 5, 6, 7, 8, 9);
 		List<Integer> sortPageAll = CollUtil.sortPageAll(2, 5, new Comparator<Integer>() {
 			@Override
 			public int compare(Integer o1, Integer o2) {
-				//反序
+				// 反序
 				return o2.compareTo(o1);
-			}}, list);
+			}
+		}, list);
+
+		Assert.assertEquals(CollUtil.newArrayList(4, 3, 2, 1), sortPageAll);
+	}
+	
+	@Test
+	public void containsAnyTest() {
+		ArrayList<Integer> list1 = CollUtil.newArrayList(1, 2, 3, 4, 5);
+		ArrayList<Integer> list2 = CollUtil.newArrayList(5, 3, 1, 9, 11);
 		
-		Assert.assertEquals(CollUtil.newArrayList(4,3,2,1), sortPageAll);
+		Assert.assertTrue(CollUtil.containsAny(list1, list2));
+	}
+
+	@Test
+	public void containsAllTest() {
+		ArrayList<Integer> list1 = CollUtil.newArrayList(1, 2, 3, 4, 5);
+		ArrayList<Integer> list2 = CollUtil.newArrayList(5, 3, 1);
+		
+		Assert.assertTrue(CollUtil.containsAll(list1, list2));
 	}
 }